Found some things when upgrading
I upgraded today from 0.4.3 to 0.5.22, had some trouble getting into my save game. These 2 changes worked for me.
It failed twice on line 214 `if global.players[player.name] then`.

It first failed on `global`, indexing a nil value. So I changed the `global = nil` definition to `global = {}`.
Then it failed on `player` being nil, I don't know when game.players is introduced, but it appears to return nil for the player in my case.
I solved it by making that `if` an `elseif`, but I have not idea if that is the correct fix.